Rework code around `Comment` object
Previously code was based around the concept of a `Codeline`. This
proved flawed when dealing with multi-line strings, since a line in a
text was not a tokenize-able string of code.

Reworking around the new `Comment` solved this, but required a complete
overhaul of the code; the new file structure and tests reflect this.